feat: moved QR component from people into common components
