+++ frontend/app/src/components/common/QR.tsx
