+++ frontend/app/src/people/auth/SignIn.tsx
@@ -5,10 +5,8 @@ import { observer } from 'mobx-react-lite';
 import { AuthProps } from 'people/interfaces';
 import { SOCKET_MSG, createSocketInstance } from 'config/socket';
 import { useStores } from '../../store';
-import { Divider } from '../../components/common';
-import IconButton from '../../components/common/IconButton2';
+import { Divider, QR, IconButton } from '../../components/common';
 import { useIsMobile } from '../../hooks';
-import QR from '../utils/QR';
 import AuthQR from './AuthQR';
 import SphinxAppLoginDeepLink from './SphinxAppLoginDeepLink';
