+++ frontend/app/src/people/widgetViews/summaries/wantedSummaries/Invoice.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import lighningDecoder from 'light-bolt11-decoder';
 import { InvoiceState } from 'people/widgetViews/organization/interface';
-import QR from 'people/utils/QR';
 import QrBar from 'people/utils/QrBar';
+import QR from '../../../../components/common/QR';
 import { calculateTimeLeft } from '../../../../helpers';
 import { QrWrap } from './style';
